integrations/agentic-frameworks/openai-agents-sdk/pyproject.toml
[project]
name = "openai-agents-example"
version = "1.0.0"
description = "OpenAI Agents SDK 集成 PPIO Agent Runtime 示例项目"
authors = [
    {name = "PPIO", email = "support@ppio.cloud"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "ppio-sandbox>=1.1.0b1",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

